lcUI cleanup and reorganisation
