+++ lcUI/widgets/clicommand.cpp
